+++ launchdarkly/resource_launchdarkly_metric_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-	ldapi "github.com/launchdarkly/api-client-go/v14"
+	ldapi "github.com/launchdarkly/api-client-go/v15"
 	"github.com/stretchr/testify/require"
 )
 
@@ -113,14 +113,14 @@ func scaffoldProjectWithExperimentationSettings(client *Client, betaClient *Clie
 	}
 
 	// Update the project's experimentation settings to make the new context available for experiments
-	expSettings := ldapi.ExperimentationSettingsPut{
+	expSettings := ldapi.RandomizationSettingsPut{
 		RandomizationUnits: randomizationUnitsInput,
 	}
-	_, _, err = betaClient.ld.ExperimentsBetaApi.PutExperimentationSettings(betaClient.ctx, projectKey).ExperimentationSettingsPut(expSettings).Execute()
+	_, _, err = betaClient.ld.ExperimentsBetaApi.PutExperimentationSettings(betaClient.ctx, projectKey).RandomizationSettingsPut(expSettings).Execute()
 	return err
 }
 
-func TestAccMetric_Basic(t *testing.T) {
+func TestAccMetric_BasicCreateAndUpdate(t *testing.T) {
 	projectKey := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	resourceName := "launchdarkly_metric.basic"
 	resource.ParallelTest(t, resource.TestCase{
@@ -148,34 +148,6 @@ func TestAccMetric_Basic(t *testing.T) {
 				ResourceName:      resourceName,
 				ImportState:       true,
 				ImportStateVerify: true,
-			},
-		},
-	})
-}
-
-func TestAccMetric_Update(t *testing.T) {
-	projectKey := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
-	resourceName := "launchdarkly_metric.basic"
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck: func() {
-			testAccPreCheck(t)
-		},
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				Config: withRandomProject(projectKey, testAccMetricBasic),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckProjectExists("launchdarkly_project.test"),
-					testAccCheckMetricExists(resourceName),
-					resource.TestCheckResourceAttr(resourceName, NAME, "Basic Metric"),
-					resource.TestCheckResourceAttr(resourceName, KEY, "basic-metric"),
-					resource.TestCheckResourceAttr(resourceName, PROJECT_KEY, projectKey),
-					resource.TestCheckResourceAttr(resourceName, KIND, "pageview"),
-					resource.TestCheckResourceAttr(resourceName, "urls.0.kind", "substring"),
-					resource.TestCheckResourceAttr(resourceName, "urls.0.substring", "foo"),
-					resource.TestCheckResourceAttr(resourceName, "urls.1.kind", "regex"),
-					resource.TestCheckResourceAttr(resourceName, "urls.1.pattern", "foo"),
-				),
 			},
 			{
 				Config: withRandomProject(projectKey, testAccMetricUpdate),